+++ src/App.tsx
@@ -10,6 +10,7 @@ import LibraryView from "./components/LibraryView";
 import CollectionsView from "./components/CollectionsView";
 import PlaylistView from "./components/PlaylistView";
 import ConfigView from "./components/ConfigView";
+import ProjectionView from "./components/ProjectionView";
 import DetailPanel from "./components/DetailPanel";
 import PlayerBar from "./components/PlayerBar";
 import AddFolderDialog from "./components/AddFolderDialog";
@@ -110,6 +111,7 @@ export default function App() {
             {view === "colecciones" && <CollectionsView />}
             {view === "lista" && <PlaylistView />}
             {view === "config" && <ConfigView />}
+            {view === "proyeccion" && <ProjectionView />}
           </main>
         </div>